Add files via upload
These two scripts are extremely simple. The config script expects your dvr ip address,  user and pass, and your desired timezone. In our setup we had an outside piece of equipment changing the time zone every day and I wasn't allowed to block them. So I set this up to run as a cron job once a minute and we never had a problem again.